chore: moved `groupExperiences` function to `utils/experience.ts`
