+++ src/app/[username]/profile-experience.tsx
@@ -3,6 +3,7 @@
 import { ExperienceRecord, ProfileAiReviewFeedback } from '@/types';
 import { useProfileAiReview } from './profile-review-provider';
 import { Bot } from 'lucide-react';
+import { groupExperiences } from '@/utils/experience';
 
 export function ProfileExperienceListWithReview({
   experiences,
@@ -129,63 +130,3 @@ const AIFeedBack = ({ feedback }: { feedback: ProfileAiReviewFeedback['experienc
     </div>
   );
 };
-
-interface ExperienceRecordGroup {
-  companyName: ExperienceRecord['companyName'];
-  companyLogo: ExperienceRecord['companyLogo'];
-  startYear: ExperienceRecord['startYear'];
-  endYear: ExperienceRecord['endYear'];
-  items: ExperienceRecord[];
-}
-/**
- * Groups an array of experience records by company name, merge continuous
- * periods of employment and separate periods with time gaps.
- * @param experiences An array of `ExperienceRecord` objects.
- * @returns An array of `ExperienceRecordGroup` objects sorted `DESC` by `startYear` property.
- */
-export function groupExperiences(experiences: ExperienceRecord[]): ExperienceRecordGroup[] {
-  if (!experiences || experiences.length === 0) {
-    return [];
-  }
-  const sortedExperiences = [...experiences];
-
-  // Sort by company name & start year
-  sortedExperiences.sort(
-    (a, b) => a.companyName.localeCompare(b.companyName) || a.startYear - b.startYear
-  );
-
-  const grouped = sortedExperiences.reduce<ExperienceRecordGroup[]>((acc, exp) => {
-    // Get most recent created group
-    const lastGroup = acc.length > 0 ? acc[acc.length - 1] : null;
-
-    // Conditions (same company & continuous employment period)
-    const isSameCompany = lastGroup && lastGroup.companyName === exp.companyName;
-    const isContinuous =
-      lastGroup && (lastGroup.endYear === null || exp.startYear <= lastGroup.endYear);
-
-    if (isSameCompany && isContinuous) {
-      // Merge to the existing group
-      lastGroup.items.push(exp);
-
-      if (exp.endYear === null) {
-        lastGroup.endYear = null;
-      } else if (lastGroup.endYear !== null) {
-        lastGroup.endYear = Math.max(lastGroup.endYear, exp.endYear);
-      }
-    } else {
-      // Add to a new group
-      acc.push({
-        companyName: exp.companyName,
-        companyLogo: exp.companyLogo,
-        startYear: exp.startYear,
-        endYear: exp.endYear,
-        items: [exp],
-      });
-    }
-
-    return acc;
-  }, []);
-
-  // Re-sort the final groups startYear (DESC order)
-  return grouped.sort((a, b) => b.startYear - a.startYear);
-}
